categoryfilter og alert

+++ src/components/InvestmentItem.js
@@ -1,9 +1,20 @@
 import React from 'react';
 import './styles.css'; 
+import Swal from 'sweetalert2'
 
 const InvestmentItem = ({ investment, deleteInvestment, editInvestment }) => {
   const handleDelete = () => {
-    deleteInvestment(investment.id);
+    Swal.fire({
+      title: 'Are you sure you want to delete this investment?',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it',
+      confirmButtonColor: '#ff0800',
+    }).then((result) => {
+      /* Read more about isConfirmed, isDenied below */
+      if (result.isConfirmed) {
+        deleteInvestment(investment.id);
+      } 
+    })
   };
 
   const handleEdit = () => {
